+++ ChangeLog.md
@@ -21,6 +21,8 @@ Current Trunk
   just changes the default, which means you can pass `-s ABORTING_MALLOC=1` to
   override the default, which was not possible before. (If you pass the flag
   and don't want that behavior, stop passing the flag.) (#11131)
+- A new `PRINTF_LONG_DOUBLE` option allows printf to print long doubles at full
+  float128 precision. (#11130)
 - `emscripten_async_queue_on_thread` has been renamed to
   `emscripten_dispatch_to_thread` which no longer implies that it is async -
   the operation is in fact only async if it is sent to another thread, while it